+++ controllers/budgettypedsctlr.js
@@ -1,8 +1,17 @@
 const budgettypeds = require('../Models/budgettypeds');
+const { recordLog } = require('./budgetauditutils');
 
 exports.addbudgettypeds = async (req, res) => {
     try {
-        const newItem = await budgettypeds.create(req.body);
+        const { username, useremail, ...bodyData } = req.body;
+        const newItem = await budgettypeds.create(bodyData);
+        
+        await recordLog({
+            username, useremail, colid: newItem.colid,
+            action: 'ADD_TYPE', module: 'BUDGET_CONFIG', recordid: newItem._id,
+            details: { budgettype: newItem.budgettype }
+        });
+
         res.status(201).json({ status: 'success', data: { item: newItem } });
     } catch (err) {
         res.status(400).json({ status: 'fail', message: err });
@@ -11,7 +20,17 @@ exports.addbudgettypeds = async (req, res) => {
 
 exports.updatebudgettypeds = async (req, res) => {
     try {
-        const item = await budgettypeds.findByIdAndUpdate(req.query.id, req.body, { new: true, runValidators: true });
+        const { username, useremail, ...updateData } = req.body;
+        const item = await budgettypeds.findByIdAndUpdate(req.query.id, updateData, { new: true, runValidators: true });
+        
+        if (item) {
+            await recordLog({
+                username, useremail, colid: item.colid,
+                action: 'UPDATE_TYPE', module: 'BUDGET_CONFIG', recordid: item._id,
+                details: { changedFields: Object.keys(updateData) }
+            });
+        }
+        
         res.status(200).json({ status: 'success', data: { item } });
     } catch (err) {
         res.status(400).json({ status: 'fail', message: err });
@@ -20,7 +39,16 @@ exports.updatebudgettypeds = async (req, res) => {
 
 exports.deletebudgettypeds = async (req, res) => {
     try {
-        await budgettypeds.findByIdAndDelete(req.query.id);
+        const { username, useremail } = req.query;
+        const item = await budgettypeds.findById(req.query.id);
+        if (item) {
+            await recordLog({
+                username, useremail, colid: item.colid,
+                action: 'DELETE_TYPE', module: 'BUDGET_CONFIG', recordid: item._id,
+                details: { budgettype: item.budgettype }
+            });
+            await budgettypeds.findByIdAndDelete(req.query.id);
+        }
         res.status(204).json({ status: 'success', data: null });
     } catch (err) {
         res.status(400).json({ status: 'fail', message: err });
